Renamed form section vals to include suffix  'val'.
This is so that the components can be given meaningful names
(to aid debugging).  Without the suffix 'val', the components were
given non meaningful name e.g. Hobby was previously named Hob
because the type for the form value was named Hobby.  Now the
type for the form value is named HobbyVal.